create type response api
