+++ src/Models/Event/EventType.ts
@@ -1,6 +1,18 @@
 export interface EventType {
   id: number
   name: string
+  slugName?: string
+  eventDate?: string
+  eventEndDate?: string
+  introduction: string
+  place?: string
+  createdAt: string
+  date?: string
+}
+
+export interface TEventType {
+  id: number
+  name: string
   slug_name?: string
   event_date?: string
   event_end_date?: string
