--- src/Models/NotificationType.ts
@@ -1,70 +1,0 @@
-export interface NotificationType {
-  id: number
-  name: string
-  slug_name?: string
-  tags?: any
-  introduction: string
-  created_at: string
-  created_time?: string
-}
-
-export const DataNotifi = [
-  {
-    id: 103,
-    name: 'LỊCH TRỰC NHẬT THÁNG 2/2023',
-    slug_name: 'lich-truc-nhat-thang-22023',
-    tags: null,
-    introduction:
-      'Nhà sạch thì mát - Bát sạch ngon cơm Văn phòng không rác thì luôn thơm, nước luôn đầy bình thì không khát',
-    created_at: '2023-01-31 16:58:06',
-    created_time: '2023-01-31 16:58',
-  },
-  {
-    id: 102,
-    name: 'LỊCH TRỰC NHẬT THÁNG 1/2023',
-    slug_name: 'lich-truc-nhat-thang-12023',
-    tags: null,
-    introduction:
-      'Nhà sạch thì mát - Bát sạch ngon cơm Văn phòng không rác thì luôn thơm, nước luôn đầy bình thì không khát',
-    created_at: '2023-01-11 15:53:19',
-    created_time: '2023-01-11 15:53',
-  },
-  {
-    id: 100,
-    name: 'HAPPY BIRTHDAY QUÝ l.2023',
-    slug_name: 'happy-birthday-quy-l2022',
-    tags: null,
-    introduction:
-      '… Mùa xuân đã đến cho trời xanh tươi hy vọng\r\nVà ngày này đã đến cho chúng ta thôi nhớ mong\r\nCòn chờ chi em hỡi ta cùng nhau đăng ký quà sinh nhật thui nào…',
-    created_at: '2023-01-03 16:31:01',
-    created_time: '2023-01-03 16:31',
-  },
-  {
-    id: 99,
-    name: 'BÀI TẬP THỂ DỤC HÀNG NGÀY',
-    slug_name: 'bai-tap-the-duc-hang-ngay',
-    tags: null,
-    introduction: 'Thể dục thể thao - Nâng cao sức khỏe',
-    created_at: '2022-12-13 11:16:39',
-    created_time: '2022-12-13 11:16',
-  },
-  {
-    id: 98,
-    name: 'LỊCH NGHỈ TẾT 2023',
-    slug_name: 'lich-nghi-tet-2023',
-    tags: null,
-    introduction: 'O Ô Ơ !!! Sắp đến Tết rồi ư :o',
-    created_at: '2022-12-07 09:26:10',
-    created_time: '2022-12-07 09:26',
-  },
-  {
-    id: 97,
-    name: 'LỊCH TRỰC NHẬT THÁNG 12/2022',
-    slug_name: 'lich-truc-nhat-thang-122022',
-    tags: null,
-    introduction:
-      'Nhà sạch thì mát - Bát sạch ngon cơm\r\nVăn phòng không rác thì luôn thơm, nước luôn đầy bình thì không khát',
-    created_at: '2022-12-05 15:46:07',
-    created_time: '2022-12-05 15:46',
-  },
-]
