+++ src/Models/Post/NotificationType.ts
@@ -1,0 +1,19 @@
+export interface NotificationType {
+  id: number
+  name: string
+  slugName?: string
+  tags?: any
+  introduction: string
+  createdAt: string
+  createdTime?: string
+}
+
+export interface TNotificationType {
+  id: number
+  name: string
+  slug_name?: string
+  tags?: any
+  introduction: string
+  created_at: string
+  created_time?: string
+}
