+++ src/Screens/Home/index.tsx
@@ -1,7 +1,7 @@
 import EventComponent from '@components/EventComponent'
 import NotificatioComponent from '@components/NotificationComponent'
-import { EventType } from '@model/EventType'
-import { NotificationType } from '@model/NotificationType'
+import { TEventType } from '@model/Event/EventType'
+import { TNotificationType } from '@model/Post/NotificationType'
 import { useGetEventQuery } from '@services/modules/event'
 import { useGetPostQuery } from '@services/modules/post'
 import commontStyle from '@styles/commont.style'
@@ -23,23 +23,23 @@ const Home = () => {
     >
       <View style={styles.homeContent}>
         <Text style={styles.contentHead}>Thông báo</Text>
-        {dataPost.data.posts.slice(0, 3).map((item: NotificationType) => (
+        {dataPost.data.posts.slice(0, 3).map((item: TNotificationType) => (
           <NotificatioComponent
             key={item.id}
             id={item.id}
             name={item.name}
             introduction={item.introduction}
-            created_at={item.created_at}
+            createdAt={item.created_at}
           />
         ))}
         <Text style={styles.contentHead}>Sự kiện</Text>
-        {dataEvent.data.events.slice(0, 3).map((item: EventType) => (
+        {dataEvent.data.events.slice(0, 3).map((item: TEventType) => (
           <EventComponent
             key={item.id}
             id={item.id}
             name={item.name}
             introduction={item.introduction}
-            created_at={item.created_at}
+            createdAt={item.created_at}
           />
         ))}
       </View>
